Objects

+++ src/brickingbad/domain/game/persistence/Save.java
@@ -1,17 +1,14 @@
 package brickingbad.domain.game.persistence;
 
-import java.util.ArrayList;
-import java.util.List;
-
 public class Save {
 
   public String name;
   public int score;
   public int lives;
-  public List<Integer> paddleCoordinates;
-  public List<List<Integer>> brickCoordinates;
-  public List<List<Integer>> ballCoordinates;
-  public List<List<Integer>> ballVelocities;
+  public List<Double> paddleCoordinates;
+  public List<List<Double>> brickCoordinates;
+  public List<List<Double>> ballCoordinates;
+  public List<List<Double>> ballVelocities;
 
   public Save() {
     paddleCoordinates = new ArrayList<>();
